Use setSoxPath to set sox path, assume 'sox'
Using `which` to determine the sox path is really confusing;
figuring out the path, if it's more than just 'sox',
should be an application-level concern.

This allows `.setSoxPath(path)` to set the sox path,
and removes util.which, an external process call,
a local cache, and a level of callbacks.
It's a win!